# New Implementation of Barabasi-Albert generator