I have rewritten the Barabasi-Albert generator to address #32.

At the same time I took the opportunity to make some speed improvements under the hood. The previous implementation, when trying to create new edges, selected a potential node uniformly at random then compared nextDouble() to the attachment probability of this node. Whilst this has a small memory footprint and works well for small graphs it has scalability issues and larger graphs would take a long time to generate as the potential nodes could be selected and rejected many times for each successful choice.

With the inclusion of WeightedChoice in the JUNG util package this was the perfect candidate for replacing the previous method and the implementation instead now works out the correct weighted probabilities of attaching to each node and then simply makes choices for the correct number of edges needed within the timestep.

The speedup is drastic as can be seen below:

Nodes |Old Impl. (sec)|New Impl. (sec)|Speedup (%)
----:|:----:|:----:|:----:
 1000 |1.13|0.99|12%
 5000 |29.56|12.20|59%
 10000 |145.40|51.16|65%
 15000 |373.50|114.25|69%
 20000 |754.69|192.91|74%
 25000 |1145.61|315.92|72%

And comes at not a great cost to memory footprint (values are eye-balled from Visual VM).

Nodes | Old Impl. Avg (mb) |New Impl. Avg (mb) | Old Impl. Max (mb) | New Impl. Max (mb) |
----:|:----:|:----:|:----:|:----:
 1000| 120 | 60 | 290 | 120
 25000| 60 | 150 | 320 | 350

The reimplementation offers drastic speedups as well as fixing the previous bug and I hope it's of an acceptable quality that it might be included!

Additionally, I have enhanced and extended the Barabasi-Albert JUnit tests to test a number of new things (graph generation for different types, rough distribution checks) to hopefully ensure further changes are compliant.