@jrtom 2016-09-04
Sorry it's taken me a while to get to this.  Looks pretty good; I'll do some polishing later (the internal data structures can be made more efficient) but it's definitely an improvement.  Thanks for enhancing the tests as well.